refactor(types): split encryption.ts into encryption-primitives.ts + response-unions.ts
