+++ packages/types/src/__tests__/server-safe.test.ts
@@ -3,7 +3,6 @@ import { describe, expectTypeOf, it, expect } from "vitest";
 import { serverSafe } from "../server-safe.js";
 
 import type {
-  AuditLogEntryServerMetadata,
   ClientAcknowledgementRequest,
   ClientBoardMessage,
   ClientChannel,
@@ -23,7 +22,6 @@ import type {
   ClientPoll,
   ClientPollVote,
   ClientRelationship,
-  ClientResponseData,
   ClientStructureEntity,
   ClientStructureEntityType,
   ClientTimerConfig,
@@ -42,21 +40,20 @@ import type {
   ServerInnerWorldRegion,
   ServerJournalEntry,
   ServerLifecycleEvent,
-  MemberServerMetadata,
   ServerMemberPhoto,
   ServerNote,
   ServerPoll,
   ServerPollVote,
   ServerRelationship,
-  ServerResponseData,
   ServerStructureEntity,
   ServerStructureEntityType,
   ServerTimerConfig,
   ServerWikiPage,
-} from "../encryption.js";
-import type { AuditLogEntry } from "../entities/audit-log-entry.js";
-import type { Member } from "../entities/member.js";
+} from "../encryption-primitives.js";
+import type { AuditLogEntry, AuditLogEntryServerMetadata } from "../entities/audit-log-entry.js";
+import type { Member, MemberServerMetadata } from "../entities/member.js";
 import type { PaginatedResult } from "../pagination.js";
+import type { ClientResponseData, ServerResponseData } from "../response-unions.js";
 import type { ServerSafe } from "../server-safe.js";
 
 // ── Count assertion ────────────────────────────────────────────────
