+++ packages/types/src/__tests__/sot-manifest.test.ts
@@ -2,13 +2,11 @@ import { describe, expectTypeOf, it } from "vitest";
 
 import type { SotEntityManifest } from "../__sot-manifest__.js";
 import type {
+  AuditLogEntry,
   AuditLogEntryServerMetadata,
   AuditLogEntryWire,
-  MemberServerMetadata,
-  MemberWire,
-} from "../encryption.js";
-import type { AuditLogEntry } from "../entities/audit-log-entry.js";
-import type { Member } from "../entities/member.js";
+} from "../entities/audit-log-entry.js";
+import type { Member, MemberServerMetadata, MemberWire } from "../entities/member.js";
 import type { Equal, Extends } from "../type-assertions.js";
 
 describe("SotEntityManifest", () => {
